IMPLEMENT SOME NEW FEATURES - FIX SEGFAULT IN BASE GUI
Implemented two of the new features of IP and Port distribution from
ticket #8, still more to do on that ticket.

Separated the redraw of the suspect list in the GUI into it's own thread with
a hard coded 5 second delay to workaround a seg fault during periods of
high suspect activity.  It probably makes sense to separate the listening
for new suspect info on sockets from GUI modification anyway.  -Ticket #87