Change GUI Tabs to a sideways orientation
Qt allows zero flexibility with the size of the tab, orientation of the tab text
and displaying a new line for longer names.  The tabs have been orientated
west and hidden behind a series of buttons that emulate the tabs behavior
by setting the tab widgets current index. The buttons allow us more
flexibility as well as the ability to specify different or additional behaviors.
- See ticket #111