Change behavior of Suspect List updates in GUI
The list no longer clears every update but instead either
a. updates the information
b. classification has changed so updates and moves the suspect
c. inserts a new suspect

As of now suspects are never removed entirely.

See ticket #112 and #113.

In respect to ticket #113, the user should only
lose a selection if the suspects classification changes, which makes sense.