Suspects sometimes aren't displayed properly in the list
When the suspect list was changed to update only a need-to-be basis, the
locks were never changed, so while it was drawing it would sometimes 
write to the suspect table without a write lock.  I fixed this and haven't been
able to reproduce the bug since.  It is likely that a flag got overwritten,
causing the suspect to show in the wrong list.  See ticket #116