Implemented Preferences Buttons - Pref errors only exit Pref Window
Implemented the Ok, Cancel, Restore Defaults and Apply buttons.

Ok - writes entered values to the config file and closes the window.

Apply - writes entered values to the config file but leaves the window open

Cancel - closes the preferences window.

Restore Defaults -  Reloads the window with the values in NOVAConfig.txt,
this will be changed once we decide on the appropriate behavior, but for
now it works like an undo button as long as you don't hit Apply or OK.

Additionally, all conditions that cause the preferences window to exit will
now only close that window allowing the main GUI window to continue to
run.