Implement Run/Stop in GUI, implement sockets for GUI to Nova
Implemented Starting and Stopping from the GUI, stop is called on exit if
nova is running, running can be done with the current configuration or using
run as which will set the run configuration values before running.  

Abstracted terminals and hid their output, you know have to  view the log
to see process output until terminals are implemented as an option. 

See Ticket #120

Stop uses sockets to send EXIT which causes the Nova processes to exit
when they can.  Socket functions are modular so other messages can be
sent easily. 
See Ticket #122