Implement main view features
- Removed old configuration views
- Implemented Main view with separate suspect list and run/stop buttons.
- Added Clear suspects button to Suspects view, this clears the suspect
hash tables to the GUI and Nova itself
- Added Exit to File menu.
- Modified Run As to read from a copy and write to the original config to
avoid re-creating the config as root. Before it moved the original and 
remade it.
- Added Hide Old Suspects to the View menu, this button clears the list of
all suspects so that only suspects with new information are displayed. 
Information on the suspects remain, this is just to remove visual clutter.
- Added Show All Suspects to the View menu, this button clears the list then
redraws all the suspects from scratch to insure that all suspects in the table
remain.
- Some non functional placehoder pages and widgets are added.
- Added a series of send functions to allow the GUI to send commands to
individual components.