Update File System Architecture
Updated all source and configurations to operate correctly using
the below changes.

Move files to relevant locations:
Read-Only Files: Nmap fingerprints and Paths are in /etc/nova/
Binaries are in	 usr/local/bin/
Logs and Scripts are in	usr/share/nova
Config/, Data/, Keys/, Templates/, Settings, Scripts are in $HOME/.nova/

Paths is a new file that contains the Path locations of installed components

Settings is a new file that will specify user settings and configurations for
parts of the Nova that don't fit into our existing categories.  This will
most likely be GUI view settings and other superficial details.

Templates is a new folder with 4 xml files that will describe the structure of
the user's honeyd configuration.

Profile describes the heirarchy and attributes of honeyd Templates.
Ports describes the attributes and behavior of ports
Nodes describes the attributes of the Haystack and Doppelganger nodes
Routes describes the topology of the Honeyd virtual network

Scripts is a new file that will specify the path to a scripts identified by name.
The user will only be allowed to use Scripts imported and tracked in this file.

Added new folder Installer into the Repository.

This folder containes all the files needed to Install Nova and three scripts:

UpdateNova: This updates all files that are maintained in eclipse source
currently,  it is like loadNovaSrc but for the installer. 
* Does not need to run as root. 
* Installed version won't be updated until InstallNova is called.

InstallNova: This creates all directories, copies all files and modifies all
permissions needed to run Nova from any directory in the terminal. Just
call NovaGUI or Individual components if needed. 
*Must be run as root. *Does not install honeyd
*May be called without removing first.

RemovaNova: This removes all directories and files created by InstallNova
to insure a clean reset if desired.
*Must be run as root. *Does not uninstall honeyd
*May be called without installing first.

Files in Installer are Nova Binaries, all honeyd's scripts, everything in ./nova
and the nmap/paths file.