Initial honeyd configuration commit, load capability, handle subnets
populated scripts, nodes,  ports, and profiles with current honeyd config.
nodes now contains subnets and the doppelganger in the group.

novagui now successfully reads from these xml files and stores config
information locally that can be modified and written later.
- See ticket # 147 

settings currently has one parameter, the current group which represents
which honeyd config preset to use.  This way the user can have multiple
configurations without having to reconfigure the entire haystack.
- See ticket # 126

subnets are handled according to a mask of the form  255.255.255.0,
masks of the form 192.168.3.0/#bits will not be stored.  For display in the
gui we can specify subnets in the 192.168.3.0/#bits form and convert back
and forth.
- See ticket # 132

All changes also fall under Ticket #133 "Configure honeyd files in GUI"