Implment Custom Serialization for Traffic Events
Event boosts serialization ran through a lot of steps to create the  archive, 
look up how to serialize the objects, serialize  and destroy the archive.  It
was the by far the largest bottleneck in the program. By serializing the
Event data as a sequence of bytes to and from a string, performance was
again drastically increased. 

I estimate it to run at about 150% the speed of using a binary archive and
~ 300% the speed using a text archive. Profiling shows that deserialization
of TrafficEvents is no longer the majority of cycles in ClassificationEngine.

When Running a T4 nmap scan of 50k ports the average
traffic was 2mb/s in and 2mb/s out. I actually had to disable the verbosity
of honeyd due to syslog prints using the majority of cycles and limiting
the speed to 800kb/s.

We may want to serialize everything ourselves for the sake of performance
over boost.