Create Nova Message Type
Created a new class called GUIMsg which is an object that is used to send
messages to nova processes from the GUI.
It is serialized manually to maintain fast communication.

GUIMsg has two members, char type, and string val. 
Type defines the action, val is the argument for those actions that need it.

val and type are private and must be accessed through functions. 
This is so that only defined types are used. 
 If the type has an arg the maximum character length of val is 255,
making the total message have a maximum size of 256 bytes. If no arg then
it is only one byte.

See ticket #128, serialization related to ticket #151