Implement google's dense_map
All tr1:unordered_maps have been converted to google dense_maps.  This
was the last of the identified performance improvements, the wiki contains
descriptions of all of them.

Speed does to seemed to have improved over the tr1::unordered_map, but
the differences in their function characteristics makes it hard to confirm in
profiling. 

Google does the work with fewer functions which makes it easier to profile,
tr1 has many functions that are distinct in the heirarchal profile and only do 
a few % each. 

Third Party testing and observation of pcapfile processing
seems to show a significant performance increase.

See Ticket #151