Implement XML Storage - Handle Subnets
We now have working persistent storage of honeyd configuration options
that should allow us to easily create complex configurations that can be
represented in the GUI. 

Because the preferences window will be redesigned to work with the new
storage and cover all of the options, the editing ports and add/clone/edit
nodes buttons have been disabled.  Editing the exisiting parameters of the
profiles has also not been tested because of the upcoming redesign.

The storage does not cover some of honeyd's more advanced options, but
the system has been designed with their implementation in mind, this will
likely be done in tandem with the redesign.

See Ticket #147

A system has been put in place to allow full subnet handling.  Full
configuration of subnets will not be included until the WSYWIG drag and
drop configuration tool for setting up network topology.  What it can do is
allow manual specification of the entire range of subnets.

See Ticket # 132