Fix for minor errors, remove profiling flag in repo makefile
Having profiling enabled by default was creating gmon.out files everywhere
and it was getting quite untidy, removed profiling as the default build, just
enable it if you want to use it.

Trailing whitespaces currently causes problems for configuration out of
NOVAConfig.txt, removed all trailing whitespaces.

Typo in Haystack was causing variables not being set.  In loadConfig, prefix
was being set to prefix[n] not prefixes[n] causing prefix to be some weird
single character or something.  With the typo fixed it works as intended.