Switch Haystack and LTM IPC from TrafficEvents to Suspects with data
In order to improve performance by sending less information between
processes, the haystack and LTM now maintain a table of unclassified
suspects.  When new information is recieved for a suspect the suspect's
featureData is created, serialized and sent to the ClassificationEngine.

Also removed uncessary libraries and took suspect and featureset out of
the ClassificationEngine namespace.