Various fixes for port knocking
The bind/listen for receiving silent alarms now uses it's local address.

The iptable rule before silent alarm is sent was allowing replies only on port
4242, this caused problems because the syn/ack was sent to port 50668, so
the rule now opens all tcp ports for the intended host temporarily.

sendto is only for connectionless sockets, switched to send since the addr is
specified in connect so for a sock_stream send is the same as sendto.

modified initial iptable rules to make all ports appear closed.