Port Knocking - Ticket #4 - Complete
Removed unecessary buffer pointer and obsolete error checking

We now move broadcasted IPTable data for the local host from the bin for
the broadcaster's addr to the bin for the local host's addr, all traffic to real
machines is supposed to be in this bin and should resolve getting different
results for IP Traffic Distribution with the same data.

IPTable is now a public member of FeatureSet for the above change.