Silent Alarm data interval bug fix
When we send out data, the time interval that data occured in is also
included for feature calculation.  Previously we would start a new interval
everytime we classified as well as when we calculated the feature set, now
we only start a new interval when we serialize the data for sending.