Modify sudoers_nova to correctly give sudo permissions to the nova group
If the sudo rule evaluated to a separate line such as %admin ALL=(ALL) ALL
because it is below the include, then the user would still be prompted for a
password or otherwise perform the behavior at that rule. By setting the
exempt_group default var for NOVACMNDS we can allow all users in the
nova group sudo permissions for NOVACMNDS without a password and it
shouldn't matter what other sudo rules are present unless they override the
defaults.