NovaConfig: Include Drop Rate configuration, port list context menu
Included Drop rate configuration of profiles using a slider bar, this value is
now saved and stored in the template files.

Added placeholding context menu actions for the port table, when the
actions are triggered they show spawn a dialog containing fields for setting
the ports number, protocol and behavior. Should this port already exist
then this dialog should be filled out with the current values.