NovaConfig: Fix several crashes related to nodes and profile dependencies
Fixed crash on saveprofile being called with a null or deleted profile.

Fixed several hash table functions to only modify the table after iteration is
completed, this prevents the iterators from being invalidated and crashing.

Fixed some functions that referenced a pointer to a HashTable object after
modifying the table, this would also sometimes cause a crash, it now copies
the object's values before modifying the table and uses the copied object.