NovaGUI: NovaConfig: Misc: Node/Profile/Subnet changes
Reverted change to QDialog from QMainwindow in NovaConfig, messing
with the inheritance at this point can cause serious unintended side effects.
Please avoid doing it lightly.

Changed the logic behind a lot of the error checking for nodes, subnets and
profiles. Put the -j flag in all projects, now it won't limit the number of jobs
it attempts to multi-thread but does not appear to create excess jobs.

Doppelganger is no longer an IP Assignment Type.