Lib: DM: Add DOPPELGANGER_INTERFACE value in NOVAConfig.txt
Included the configuration value DOPPELGANGER_INTERFACE to specify
the loopback interface on which the Doppelganger should be placed.
Although this value is not currently used, in the future we will be able to use
it for the use-case(s) when there is more than one loopback interface, no
loopback interface or the loopback isn't named lo (ubuntu default).