Remove global settings vars and read from NOVAConfiguration file setters
Having duplicate variables all over after the merge was a problem. This
eliminates them and uses a global NOVAConfiguration object to read in 
the config file a single time and then use those vars across all states.

Right now this has no locking, but that could easily be added later if we
decide to actually edit the stuff inside the NovaConfiguration class.