Make initUserConfigs not exit in case of failure
The call to gksudo fails in Backtrack 5, but the installation works and runs fine 
otherwise. Rather than forcing an exit upon Init failing, just log the error and
try and continue. In general it's probably a better idea to keep running rather
than force an exit if there's an error.