CE: DM: Novad: Integrate DM, moved dm functionality to classification.
Changed CE Silent Alarm logic to call when classification changes or suspect is
hostile.  Network silent alarms occur when there is new information and the above
occurs, DM rules are added if dm is enabled and the above occurs. This only
applies to suspects that are from live capture, not pcap files.