Convert NOVAConfiguration to a singleton
The configuration is used everywhere and there should only be one instance
per process. This ensures there can only be one, and also gets rid of the 
hassle of either passing a pointer to a global version of it all over or trying to
reread the configuration file and ending up with out of sync config objects. 
Also renamed to Config (shorter, NovaConfig was taken, and there's really
no need for things to be prefixed with Nova to begin with)