Fix "double free or memory corruption" error
Featureset contained a pointer to another featureset for the unsent SA data
, this was causing problems since it wasn't being deep copied correctly and 
it was getting freed more than once when two featureSets that pointed to 
the same unsent data featureset were deleted. 

It was a bad idea to have a pointer to a class of it's own type anyway,
basically a recursize definition. Moved the unsent FeatureSet to the suspect
class instead.