Move main() for Unit Testing
In order to do Unit Tests on ClassificationEngine.cpp and any other classes
inside of Novad, they need to be linked to the test executable or built with
it. Because Novad is not a static or dynamic library, the only way to do this is
to actually include the same source files during the build of the test 
executeable. However, this had the problem of having "multiple definition of main()". I tried taking Novad.cpp out of the list of things to build, but a huge chunk of functionality is still inside Novad.cpp, and files such as Control.cpp need it. Furthermore, if I want to test the numerous functions in Novad.cpp,
I also need to include that file in the test executable build. 

TLDR summary: Moved main() to it's own file, I had good reasons.

In the future we should really take all the functions of Novad and put them 
into classes or at least their own files grouped by functionality. For now, 
this is a stop gap solution so I can keep building up the Unit Tests.