NOVAD: Fix Ann point null problem
It turns out it wasn't directly a problem with SuspectTable, but how it
was used. The CE used to lock the table, loop once to update the evidence
and then loop a second time to update the classifications. However, after
the Great Refator, the CE could no longer lock the entire table. Instead, it
was just locking the suspects one at a time. If a suspect arrived between
the CalculateFeatures loop but before the Classify loop, the Ann point
 would be null and a segfault would happen.

Fixed by updating the evidence and classification of a suspect at the same
time before checking it back into the table. NeedsFeatureUpdate became
unused in the process, so removed it from Suspect.

Fixes #94