Lib: Simplify LOG Macro
all log calls automatically include the file and line number, the first argument is
the level, the second is the basic string and the third is the advanced string.
Don't specify the line and file anymore, if you want the same message for both,
pass an empty string for the third argument like so: 

LOG(INFO, "Same message", "");

which would output 'Same message' to lib notify and 
'File <file> at Line <line>:  Same message' in the logs.