BUILD: System() return value problem
There were three places in which there were system() calls made that did 
not use the return value at all, which throws a compiler warning after 
building. Went in, encased the system() calls in if statements, and if they 
don't return 0 (a semi-standard return value for success for the system() 
call), log and continue.


Tickets:
Issue #53