MISC: Creating the email alert script, cron script, and integrating into install
Created two scripts, nova_mailer and NovaCron that perform the email
alert loop and adding the script to cron, respectively. They both have 
optional parameters for changing the location of the named pipe and the
location of the file containing the emails; for now, the emails path works 
great, but there are some considerations to take into account that need to 
be researched more (changing the 40-nova.conf file to reflect new pipe
location, etc.). That will just be left to the user for now.

Tickets:
Issue #34