MISC: Adding mail alert script and cron add script.
Created a script that reads from a named pipe specifically populated with
emails from Nova, and then sends them out after the read loop is done. 
The other script is set up to manage the cron status of the mailer script;
it will add the script to the crontab if it's not there, delete it if given the 
-d flag, and update the entry when given a -u flag and two absolute paths:
one for the pipe and one for the emails file.

Tickets:
Issue #34