MISC: Quick fix for mailer script
Had to break the read loop for the pipe; before, if there was nothing in the 
pipe, the process would halt on the read. Then, if a minute goes by, there 
would be another process started by cron...etc. It was messy. Now, if at the 
time of the read loop there isn't anything to read, it'll kick out and wait until 
the next loop.

Ticket:
Issue #34