LIB: Overloading SetUserPreferences for single service changes
Went into the Logger class and added an overloaded version of 
SetUserLogPreferences to accept a Nova::Service, Nova::Levels, and char
argument to change just one service. Up to this point it will modify the
preference string and update the Config class, as well as the
m_messageInfo Logger class member. Working on getting it to write out 
to the Config file.