LIB: SetUserLogPreferences for specific services
Overloaded the SetUserLogPreferences method to accept a different set 
of arguments (a Nova::Service, a Nova::Level, and a character representing
a modifier for range) that will allow finer control over log preferences at 
any given time. It will update the struct in the Logger class and the string
in the Config class with the new service:level selections.

No Tickets