LIB && NOVAD: Allow for dynamic serialization
Added methods to procure the length that the suspect will be 
when using either the SerializeSuspect() method or 
SerializeSuspectWithData() method. This allows the buffer 
that gets passed to be initialized to the exact length required
for that suspect in that situation, obviating the need for a 
static 65536 - n sized buffer. 

Tickets:
Issue #70