MULT: Modification of Suspect Serialization methods to consolidate the different calls into one call with multiple arguments
Modified the Serialize and Deserialize methods for Suspects to allow a 
single call to do the work of both the vanilla and the WithData versions.
Did some preliminary work on Issue 71, but I need to find a way to do it 
more elegantly and centralized. Gonna be tricky to test it though.

Tickets:
Issue #71