Novad: Lib: Major Performance Enhancements
Number of performance enhancements including:
- const ref usage during parameter passing to decrease allocation calls
- Removal of unecessary workflow
- Removal of TCP session table
- Removal of evidence vector
- Removal of m_haystackEvents variable

As packets come in they are created only once, are read directly
by the target featureset, the data extracted and the packet deallocated.

During classification the only thing left to do is calculate the enabled
features and then classify on that feature set.

As of this commit novad imposes no noticeable bottleneck on a local host.

Unfortunately the version of honeyd we are currently using is making
excessive use of the rsyslogd process so much so that even a doppelganger
and no haystack is enough to throttle bandwidth and max cpu usage during
an "nmap -sS -p 1-65535 <local net>".