Build: Novad: Check pcap stats during classification, debug gdb pcap perm
Included a sudoers file in the debug installation that provides the gdb with
no password prompt permissions when executed with sudo. This allows us
to debug novad in eclipse as a normal user by simply installing the debug
build and configuring the eclipse workspace debug settings to run
'sudo gdb' instead of simply 'gdb'.

Moved packet capture stats to the classification loop - Checking the stats
everytime we get a packet was overkill and actually caused packets to be
dropped.  By only checking the stats during the classification process the
additional overhead is mitigated and so far we have yet to see any dropped
packets, we should consider lowering the buffer size and snaplen
for further gains in the packet handler.