UI_WEB: Fixing some interim problems with editHoneydNodes
Fixed the js and the UI_Core side to reflect the correct case for when a node
is added that belongs to no subnet in the subnet table. Before, in UI_Core
the HoneydConfig method for deleting a node would slip a return true 
through the cracks, but not anymore.