UI_WEB: Input Validation for the Web UI
Added javascript to validate the non-string fields of the /configNova page;
if an invalid value is found, the page will redirect to the error page which will
print the list of errors that were aggragted through iteration of the input
fields. Still have to look at the other pages to see what I can do there but 
config should be done for the most part.