LIB, NOVAD, QT_GUI: Adding SMTP functionality
Added programmatic SMTP functionality to the Logger, so when the mail
message is called an SMTP message is constructed and sent using libcurl.
If you do not set curl in your linking options for Novad and 
don't add it to the CLI then Nova won't compile. Just add -lcurl and it should 
take care of the rest, I updated the .pro file for convenience but you MUST
do the other ones.

Also, updated the dependencies list on the wiki to reflect the need for 
libcurl3-dev and libcurl4-gnutls-dev. Though we only use SSL right now, for
some reason the libcurl4-openssl-dev and libcurl4-gnutls-dev libraries are 
mutually exclusive, yet the gnutls-dev one seems to subsume the openssl
one and contains the same functionality, with the added advantage of TLS
support should I ever get it work correctly.

Tickets:
Issue #184