ALL: Add a percent of haystack contacted feature
Still need to add training data for it. Going to wait on pushing that until
I've got some other feature/classification changes working as well. 

There were some misc changes to the feature sets in order to get this to
work. We weren't really using the subtraction operator of the featureset
and I believe the new implementation without all of the adding and
subtracting of featuresets performs better. The sent and unsent 
featuresets both get evidence added to them, and the "sent" featurest (
m_features) has all of the data while the unsent one is cleared every SA 
send like it was before. 

In an abstract algebra sence, featureSet was switched from an algabric 
group to a monoid.