Qt: Fix profile renaming bug, update and minor tweak to novaconfig.ui
Fix prof rename bug:  We were previously assuming an item for a node based list would exist for a profile, however this is not the case for unused
profiles, we now check for NULL before updating, if the value is null no action is needed.

Update default indicies for novaconfig, we now open to the general 
settings page  and when viewing profiles we view the general profile options by default instead of the ports and services tabs.

Added a few spacers to make novaconfig look  cleaner and more professional.