HHCONFIG: Getting the script table up and running
Right now the ScriptsTable object will pull all the scripts from the 
Script_Table in HoneydConfiguration into a table designed for matching
port services and osclasses to scripts. Haven't implemented any of the 
actual matching thus far, was trying to find a good place to do it. I suspect
either during profile generation or after the Personality Tree structure is
completed (most probably the latter) would be chocie places to add such
functionality. 
Also, renamed ScriptTable to ScriptsTable to avoid a naming conflict 
with the ScriptTable hashmap in the NovaGuiTypes; did the same for 
the PortsTable map that's in AutoConfigHashMaps.h.