***commenting***
commented out code for scripttable stuff right now; it'll crash the novagui 
because it changes the port names in the ports table without maintaining
the open port that's there; because of this, when the GUI tries to load the 
profiles, it accesses the empty key a bunch for populating the profile tab
(in the preferences) and then segfaults. The commented code works, but 
needs to make companion updates to the profile table which I haven't
figured out a way to do yet. Just leave it as it is and I'll take care of it on 
Monday.