HHCONFIG: Getting ScriptTable integrated
Just need to either have GetProfiles return the correct pointer to the
profile, and not a copy, or access the profile directly and modify
the ports that way. Then test to see whether the ports subtree in 
the profiles.xml file actually relfect those changes.

Currently, AddAllPorts will add both a port that contains the correct
script/behavior, etc., as well as an open version in case some of the
premade profiles have that set up prior to the scanning. Once we start
calling CleanPorts the need for that may go away, but for now, better
safe than sorry.