UI_QT: Fix bug, initial connection to Novad would timeout if lots of suspects
The uptime thread was calling IsNovaUp when the other thread was
still in the process of fetching all of the suspects. The combination was
causing timeouts for some reason and making you need to attempt to
connect multiple times before it would work.

Also made it so it wouldn't draw the suspects until they've all been sent.