HHCONFIG: Finish ScriptTable stuff and fix some ptree errors
The ptree tags in the profiles got messed up during the Great Refactor so
fixed those.

Added two new scripts to the scripts.xml. Put in a specific Linux telnet
script as the only two previously available were a Windows version 
(which uses the Linux script anyways) and one for routers (which is under
type embedded; I'd add that to the osclass tag, but without having a 
version for it I fear it would require quite a bit more token matching to
get the same results. Will look more into this later.).

ScriptTables seem to be done; need to test it a bit more, make sure that 
the right scripts come out for a scanned windows machine, etc. But as far
as it looks right now, it seems like the functionality is all there, just may
need to be tweaked for some corner-y cases.