HHCONFIG: Adding avg port count and node tab
Added a tab to the profiles page in the NovaConfig dialogue for the QT 
GUI. Once finished, it will provide a list of nodes that use the profile that is selected, as well as any nodes that use child profiles of the selected profile.

Added Avg Port Count for use in determining how many ports to give to 
generated nodes in the auto config tool.