HHCONFIG && UI_QT: Column resizing for nodes tab and redoing a fix
Added column resizing so that after a new profile is selected the columns 
will automatically resize to a preferred width s.t. all items are clearly 
visible (no ellipses). 

Also had to re-add a change to conform with the coding standard, it had 
half of the fix in the repo and the other half wasn't for some reason.