HHCONFIG && UI_CORE && UI_QT: Adding new fields to NodeProfile and fixing a logging thing
Added a LOG call where there was once a cout, and added the m_nodeKeys 
variable to NodeProfile, which is a list of Nodes associated with the given 
profile. In addition, made the second element in the m_ports pair another
pair that contains a bool (which was once the second on its own) 
representing inheritance and a double, representing a distribution value.
Thus far everything compiles and runs correctly.