UI_WEB: Multiple Interface capability
Added multiple interfaces (but NOT multiple loopbacks yet) into the 
basic Options page; it'll save the selected interfaces correctly and modify
the m_ifIsDefault variable correctly now. There was some javascript/cvv8
madness that was preventing the conversion from a javascript boolean to 
a c++ boolean that was messing things up. Going to test out a couple more
things, but for now it compiles and will work in most cases. Call this an
interim commit for now.