HHCONFIG: Adding command line arguments
(REQUIRED) 
-n NUM_NODES: should come first in the ordering; NUM_NODES is an
integer greater than 0. If all of the characters in the argv string at the 
"-n" + 1 index are not digits (i.e. "3.14", "-10", etc.) execution stops.

(OPTIONAL)
-f NMAP_SCAN_TO_PARSE: if you don't want to scan your interface's
subnets, but instead want to generate profiles and nodes from an already
completed scan, throw in the -f flag with an accompanying Nmap 6 
generated XML output file. NMAP 6 REQUIRED. Cannot be used with the 
-a command line flag.

-a ADD_SUBNET1, ADD_SUBNET2,...: if you want to generate profiles from
more subnets than just the ones your interface(s) is/are connected to, use
the -a flag. Must be a comma delimited list, no spaces around the commas. 
ADD_SUBNET* should be a string of the form XXX.XXX.XXX.XXX/24. 
Note that, currently, execution will hang if you add subnets that nmap 
cannot scan (don't exist, wrong address/bitmask, etc.). Will work on fixing
that later. Cannot be used in conjunction with the -f flag.