UI_WEB: Integrating honeydhostconfig with web_ui
Took the first steps towards integrate autoconfiguration ability with the 
web UI. At present, it allows to user to designate a number of nodes to 
generate, as well as selecting additional subnets to pass to the 
executable. No support for specific nmap XML outputs yet, though
the functionality exists within the main method for the class.  

It generates the correct string for the system call, and that's where my 
testing stopped. I will be testing actual generation of profiles and 
displaying of the generated profiles and nodes tomorrow, figured this
would be a good stopping point.