HHCONFIG + UI_WEB: More Web UI integration
Got it to be properly making the calls to the executable, redirecting when 
necessary, etc. Need to fix a few things in HHCONFIG, mainly a bit
of overlogging occurring due to some things getting hit that shouldn't
be, as well as hhconfig hanging when either a subnet was scanned that
doesn't exist/you don't have access to OR when the AutoConfig has 
already run.