UI_Core: Revert HoneydConfiguration changes that broke ptrees
In delete node there was a check for a default node that doesn't exist, this was removed.

In UpdateProfileTree a conditional was added that would cause the function to stop just before updating the root node, thus the written xml was never correct and full of junk. This has been reverted.