UI_WEB: Setting up more autoconfig stuff
Right now, since the autoconfig tool dies if there's more than one subtree
of osclasses found in the scans, can't really prepare for any post-utility-run
review of information; however, added interface selection into the web UI
and hooked it in to the code behind to allow for complete user command 
line immersion. Still need to write a JS function to ensure that at least 
one subnet is added or one interface is selected, but that should mainly
be cut-paste-and-modify from the basicConfig jade family.