MULT: Super commit for EoD June 29
Modified profiles.xml to have the "generated" bool tag, denoting whether
it was created by HHCONFIG or not. Made install-hhconfig-* move the 
sudoers file so that the password prompt doesn't come up for the 
web UI.

Added the m_generated member variable to NodeProfile struct, and 
modified all requisite methods to conform to having this new value,
including writing and reading ptrees, etc.

Did more work on the Web UI as well, still having some trouble 
populating the m_nodeKeys vector for the NodeProfiles after a 
LoadAllTemplates, started working on that but didn't debug that much
yet, figured I'd save it for Monday. Aside from that, the Web UI 
integration is almost completed (I promise hahaha) just needs a better
representation of what the AutoConfig tool actually did to show the user.
Ideally it'll be simpler than the Node or Profiles pages for the Haystack
but it may be out of my hands.