UI_WEB && HHCONFIG: Exposing results of scan/node generation, a bugfixing story
Inner for loop for a printing function was incrementing the variable of the 
outer for loop, leading to reliable segfaults during web UI post-scan
function calls. Fixed that, also added a function to ensure that there's 
at least one interface selected or one subnet added before submitting
the POST in hhautoconfig.jade, essentially some input validation stuff.

Everything works pretty nicely between the HHCONFIG executable and the
web UI, there's really only the case of attempting to scan and parse a 
subnet that doesn't exit/you don't have access to left to account for 
(probably easy fix) and the case for allowing multiple, sequential uses
of the autoconfig tool, which we currently don't (much harder).