UI_WEB: Fixing first run review failure
On the first run of the autoconfig scan, the profiles would be outputted
correctly on the review page but the nodes remained empty. Added a 
call to LoadNodesTemplate after the group gets changed to construct the
correct NodesTable and thus output the correct nodes on the page. 

Also changed the button class for the Haystack Autoconfiguration button 
to maintain uniform appearance and functionality with the rest of the 
buttons. 

Changed the empty array constructor for the arrays in the app.get for 
/autoConfig to calls to new Array() from the previous calls to [].