UI_CORE: Fixing semi-colon missing error in VendorMacDb
The GenerateRandomMac(string) method would sometimes return MAC
addresses that were properly generated, with one exception: the 
semicolon separating the three hex prefix from the rest of the address 
would be missing. Upon further investigation, it was found that 
after adding the prefix in its entirety to the addrStrm variable, if the 
sprintf to put addr.second into addrBuffer returned a character array
with an odd number of elements, you would see the semicolon go missing
because the increment operation on i wouldn't get hit, thus disallowing 
the semicolon conditional to execute. I would assume that this would 
happen with any odd-length testStr, so I added a check for that in the 
conditional, and have not seen the problem pop up since.

Also did some formatting stuff, people keep crunching everything together