UI_WEB: Add slider bars to represent distribution data
Added some HTML5 slider bars into editHoneydProfiles, s.t. if you elect
to inherit the MAC vendor from the parent profile and the parent profile
has multiple vendors, it will create a three tag element for each one, 
showing the name, the slider bar, and a label to show the current 
distribution values among the n elements. Put in some rudimentary code
for maintaining a sum of 100 across the distributions when the slider bars
are used, but it's got bugs. Gonna fix it and then push.